+++ src/main/kotlin/org/kalasim/Resource.kt
@@ -226,3 +226,6 @@ class ResourceStatistics(resource: Resource) : Jsonable() {
         "occupancy" to occupancy.toJson()
     }
 }
+
+
+