+++ src/main/kotlin/org/kalasim/misc/Util.kt
@@ -157,6 +157,6 @@ fun <T> repeat(n:Int, builder: (Int) -> T) = (1..n).map{ builder(it)}
 
 
 // https://stackoverflow.com/questions/48007311/how-do-i-infinitely-repeat-a-sequence-in-kotlin
-fun <T> Iterable<T>.repeat() = sequence<T> {
+fun <T> Iterable<T>.repeat() = sequence {
     while (true) yieldAll(this@repeat)
 }