+++ src/test/kotlin/org/kalasim/examples/api/Cars.kts
@@ -11,7 +11,7 @@ class Car : Component() {
 
     override fun process() = sequence {
         request(driver) {
-            hold(1.0)
+            hold(1.0, description = "driving")
 
             wait(trafficLight, "green")
         }