+++ src/test/kotlin/org/kalasim/examples/api/ResourceSelectionClassic.kts
@@ -1,0 +1,21 @@
+//ResourceSelection.kts
+import org.kalasim.Component
+import org.kalasim.Resource
+import org.kalasim.createSimulation
+import org.kalasim.misc.repeat
+
+createSimulation {
+    val resources = List(3) { Resource() }
+
+    // we could also mimic round-robin with infinite stream
+    val resIter = resources.repeat().iterator()
+
+    object : Component() {
+        override fun process() = sequence {
+            val r = resIter.next()
+            request(r) {
+                hold(10)
+            }
+        }
+    }
+}